requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
datasets>=2.18.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
trl>=0.8.0

# Experiment tracking
wandb>=0.16.0

# Evaluation
bert-score>=0.3.13
rouge-score>=0.1.2
openai>=1.14.0

# Data pipeline
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
tqdm>=4.66.0

# Inference server
fastapi>=0.110.0
uvicorn>=0.29.0
sse-starlette>=2.0.0

# Utils
python-dotenv>=1.0.0
pyyaml>=6.0
